refactor: fix file name and change dummy data
